moved gui scripts to their own dir
