+++ src/AV_Spex/av_spex_the_file.py
@@ -30,8 +30,8 @@ class LazyGUILoader:
         if cls._QApplication is None:
             from PyQt6.QtWidgets import QApplication
             # Update imports to use the new UI modules
-            from .utils.gui_checks_window import ChecksWindow
-            from .utils.gui_main_window import MainWindow
+            from .gui.gui_checks_window import ChecksWindow
+            from .gui.gui_main_window import MainWindow
             cls._QApplication = QApplication
             cls._ChecksWindow = ChecksWindow
             cls._MainWindow = MainWindow
